requirements.txt
aiogram==3.17.0
aiohttp==3.11.13
aiosqlite==0.21.0
python-dotenv==1.0.1
